fsmc_nand.c: Fixed data abort inside change_bit()
Since change_bit() requires a (unsigned int *) as second arg,
the correct definition of err_idx[] array declared as
local variable of fsmc_correct_data() is the following:

	u32 err_idx[8];

+++ drivers/mtd/nand/fsmc_nand.c
@@ -581,7 +581,7 @@ static int fsmc_correct_data(struct mtd_info *mtd, uint8_t *dat,
 					struct fsmc_nand_data, mtd);
 	struct fsmc_regs *regs = host->regs_va;
 	unsigned int bank = host->bank;
-	uint16_t err_idx[8];
+	uint32_t err_idx[8];
 	uint64_t ecc_data[2];
 	uint32_t num_err, i;
